+++ src/Cache.php
@@ -13,6 +13,10 @@ class Cache
 
 	public function __construct($basePath, $cleanupInterval = 3600)
 	{
+		if (!is_int($cleanupInterval) && false !== $cleanupInterval) {
+			throw new \Mpdf\MpdfException('Cache cleanup interval has to be an integer or false');
+		}
+
 		if (!$this->createBasePath($basePath)) {
 			throw new \Mpdf\MpdfException(sprintf('Temporary files directory "%s" is not writable', $basePath));
 		}
@@ -106,7 +110,9 @@ class Cache
 
 	private function isOld(DirectoryIterator $item)
 	{
-		return $item->getMTime() + $this->cleanupInterval < time();
+		return $this->cleanupInterval
+			? $item->getMTime() + $this->cleanupInterval < time()
+			: false;
 	}
 
 	public function isDotFile(DirectoryIterator $item)