+++ src/ServiceFactory.php
@@ -56,8 +56,10 @@ class ServiceFactory
 
 		$tableOfContents = new TableOfContents($mpdf, $sizeConverter);
 
-		$cache = new Cache($config['tempDir']);
-		$fontCache = new FontCache(new Cache($config['tempDir'] . '/ttfontdata'));
+		$cacheBasePath = $config['tempDir'] . '/mpdf';
+
+		$cache = new Cache($cacheBasePath, $config['cacheCleanupInterval']);
+		$fontCache = new FontCache(new Cache($cacheBasePath . '/ttfontdata', $config['cacheCleanupInterval']));
 
 		$fontFileFinder = new FontFileFinder($config['fontDir']);