+++ CHANGELOG.md
@@ -24,6 +24,8 @@ mPDF 8.0.x
 * Skip non supported wrappers when resolving paths (#1204, @MarkVaughn)
 * Fixed SVGs using a style tag, has styles ignored ( Requires ext-dom ) (#450, @antman3351)
 * Allows `{nb}`, `{nbpg}`, `{PAGENO}` and `{DATE ...}` substitution in body (#172 and #267, @Dasc3er)
+* Cache now creates a dedicated subdirectory `/mpdf`.
+* It is possible to disable automatic cache cleanup with `cacheCleanupInterval` config variable
 
 mPDF 8.0.0
 ===========================